+++ _CORE_PROJECTS/HaleyExtensions/Extensions/ObjectConversion.cs
@@ -36,28 +36,6 @@ namespace Haley.Utils
         {
             return (T)_changeType(value, typeof(T));
         }
-        public static XElement ToXml(this object source)
-        {
-            Type _type = source.GetType();
-
-            #region Abandoned - To Consider interfaces
-            ////If the source has any Interface properties, we just get them as extratypes.
-            //Type[] extraTypes = _type.GetProperties()
-            //    .Where(p => p.PropertyType.IsInterface)
-            //    .Select(p => p.GetValue(source, null).GetType())
-            //    .ToArray();
-
-            //DataContractSerializer serializer = new DataContractSerializer(_type, extraTypes);
-            //serializer.WriteObject(xw, source);
-            #endregion
-            XmlSerializer serializer = new XmlSerializer(_type);
-            //TO IGNORE UNWANTED NAMESPACES
-            XmlSerializerNamespaces ns = new XmlSerializerNamespaces(new[] { XmlQualifiedName.Empty });
-            StringWriter sw = new StringWriter();
-            XmlTextWriter xw = new XmlTextWriter(sw);
-            serializer.Serialize(xw, source, ns);
-            return XElement.Parse(sw.ToString());
-        }
         #endregion
 
         #region Helpers
