+++ _CORE_PROJECTS/HaleyExtensions/HaleyExtensions.csproj
@@ -14,15 +14,15 @@
 		  <IncludeSymbols>true</IncludeSymbols>
 		  <SymbolPackageFormat>snupkg</SymbolPackageFormat>
 		  <PackageIcon>Haley.png</PackageIcon>
-		  <PackageReleaseNotes>Added byte and string helpers</PackageReleaseNotes>
+		  <PackageReleaseNotes>Added serializations helpers.</PackageReleaseNotes>
 		  <RepositoryUrl>https://github.com/rmsmech/HaleyProject</RepositoryUrl>
 		  <RepositoryType>GIT</RepositoryType>
 		  <SignAssembly>true</SignAssembly>
 		  <AssemblyOriginatorKeyFile>HaleyProject.snk</AssemblyOriginatorKeyFile>
-		  <AssemblyVersion>1.0.5.0</AssemblyVersion>
-		  <FileVersion>1.0.5.0</FileVersion>
+		  <AssemblyVersion>1.0.6.0</AssemblyVersion>
+		  <FileVersion>1.0.6.0</FileVersion>
 		  <PackageTags>IOC, DependencyInjection, DI</PackageTags>
-		  <Version>1.0.5</Version>
+		  <Version>1.0.6</Version>
 	</PropertyGroup>
 	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
 		<DebugType>portable</DebugType>
